script/codefresh-argo-workflows/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.1
garage-admin-sdk-python==0.9.0
google-auth==2.28.1
idna==3.6
kubernetes==29.0.0
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==2.2.1
websocket-client==1.7.0